Move FarmReader onnx test back into test_reader.py
